grocery_route_optimizer/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.15.0
webdriver-manager>=4.0.1

# GUI dependencies
# tkinter comes with Python by default
Pillow>=10.1.0

# Data processing
pandas>=2.1.3
numpy>=1.25.2

# Geolocation and mapping
geopy>=2.4.0
folium>=0.15.0
networkx>=3.2.1

# Web scraping helpers
fake-useragent>=1.4.0
urllib3>=2.1.0

# JSON and API handling
jsonschema>=4.20.0

# Path finding algorithms
python-tsp>=0.3.1

# Development tools
python-dotenv>=1.0.0